General refactor and clang-tidy cleanup

+++ include/sffdn/matrix_gallery.h
@@ -21,15 +21,15 @@ enum class ScalarMatrixType : uint8_t
 };
 
 // Generates a square matrix of size N x N based on the specified type.
-std::vector<float> GenerateMatrix(size_t N, ScalarMatrixType type, uint32_t seed = 0);
+std::vector<float> GenerateMatrix(uint32_t N, ScalarMatrixType type, uint32_t seed = 0);
 
-std::vector<float> NestedAllpassMatrix(size_t N, uint32_t seed = 0, std::span<float> input_gains = std::span<float>(),
+std::vector<float> NestedAllpassMatrix(uint32_t N, uint32_t seed = 0, std::span<float> input_gains = std::span<float>(),
                                        std::span<float> output_gains = std::span<float>());
 
 struct CascadedFeedbackMatrixInfo
 {
-    size_t N;                     // Number of channels
-    size_t K;                     // Number of stages
+    uint32_t N;                   // Number of channels
+    uint32_t K;                   // Number of stages
     std::vector<uint32_t> delays; // Delays, size: (K - 1) x N
     std::vector<float> matrices;  // Feedback matrices, size: K x N x N
 };
@@ -43,7 +43,7 @@ struct CascadedFeedbackMatrixInfo
  * @param gain_per_samples Gain per sample (default: 1.0)
  * @return CascadedFeedbackMatrixInfo
  */
-CascadedFeedbackMatrixInfo ConstructCascadedFeedbackMatrix(size_t N, size_t K, float sparsity, ScalarMatrixType type,
-                                                           float gain_per_samples = 1.f);
+CascadedFeedbackMatrixInfo ConstructCascadedFeedbackMatrix(uint32_t N, uint32_t K, float sparsity,
+                                                           ScalarMatrixType type, float gain_per_samples = 1.f);
 
 } // namespace sfFDN
