+++ src/matrix_gallery_internal.h
@@ -7,11 +7,11 @@
 
 namespace sfFDN
 {
-Eigen::MatrixXf RandN(size_t N, uint32_t seed = 0);
-Eigen::MatrixXf RandomOrthogonal(size_t N, uint32_t seed = 0);
+Eigen::MatrixXf RandN(uint32_t N, uint32_t seed = 0);
+Eigen::MatrixXf RandomOrthogonal(uint32_t N, uint32_t seed = 0);
 Eigen::MatrixXf HouseholderMatrix(Eigen::VectorXf v);
-Eigen::MatrixXf RandomHouseholder(size_t N, uint32_t seed = 0);
-Eigen::MatrixXf HadamardMatrix(size_t N);
-Eigen::MatrixXf CirculantMatrix(size_t N, uint32_t seed = 0);
-Eigen::MatrixXf AllpassMatrix(size_t N, uint32_t seed = 0);
+Eigen::MatrixXf RandomHouseholder(uint32_t N, uint32_t seed = 0);
+Eigen::MatrixXf HadamardMatrix(uint32_t N);
+Eigen::MatrixXf CirculantMatrix(uint32_t N, uint32_t seed = 0);
+Eigen::MatrixXf AllpassMatrix(uint32_t N, uint32_t seed = 0);
 } // namespace sfFDN
